requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.4
Django==3.1.3
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.6.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.4
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
uritemplate==3.0.1
wrapt==1.12.1
django-ckeditor~=6.0.0

django-mptt~=0.11.0
